Try to avoid more class loading issues... Again
