+++ src/main/java/io/github/crucible/bootstrap/CrucibleServerMainHook.java
@@ -103,16 +103,4 @@ public class CrucibleServerMainHook {
         System.setOut(originalOut);
         System.setErr(originalErr);
     }
-
-    // Too lazy for a coremod
-    public static void coremodHandleLaunch(File mcDir, LaunchClassLoader classLoader, FMLTweaker tweaker) {
-        classLoader.addClassLoaderExclusion("io.github.crucible.bootstrap.");
-        try {
-            // Ensure our config is loaded way before everything that may need it
-            Class.forName("io.github.crucible.CrucibleConfigs", true, classLoader);
-        } catch (ClassNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-        Lwjgl3ifyGlue.doCoremodWork(classLoader);
-    }
 }
